backend/pyproject.toml
[project]
name = "simple-social-backend"
version = "0.1.0"
description = "Backend API for Simple Social"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.121.0",
  "sqlmodel>=0.0.27",
  "uvicorn>=0.38.0",
  "psycopg[binary]",
  "pika>=1.3.2",
  "python-multipart>=0.0.9",
  "python-dotenv",
  "requests",
  "httpx",
  "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
]

# deine CLI-Kommandos – jetzt mit NEUEM Paketnamen
[project.scripts]
social-seed = "simple_social_backend.cli:seed"
social-api  = "simple_social_backend.cli:start_api"

# für src-Layout mit Hatchling: sag ihm, welches Paket gebaut werden soll
[tool.hatch.build.targets.wheel]
packages = ["src/simple_social_backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
